api/requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
elasticsearch_serverless
requests==2.32.0
python-dotenv==1.0.1
flask-cors==4.0.1
eventregistry==9.1